Improve: Extract example to directory
